requirements.txt
py-cord
motor
python-dotenv
